requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
Deprecated==1.2.15
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
kubernetes==31.0.0
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
posthog==3.7.3
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybaseball==2.2.7
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.10.0
PyNaCl==1.5.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.4
tomli==2.1.0
tqdm==4.67.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
zipp==3.21.0
